Create Find Immediate Smaller Than X
Given an array arr[] of size N containing positive integers and an integer X. You need to find the value in the array which is smaller than X and closest to it ( if no such element exists the answer should be -1)

Example 1:

Input:
N = 5
arr[] = {4 67 13 12 15}
X = 16
Output: 15
Explanation: For a given value 16, there
are four values which are smaller than
it. But 15 is the number which is smaller
and closest to it with minimum difference
of 1.
Example 2:

Input:
N = 5
arr[] = {1 2 3 4 5}
X = 1
Output: -1
Explanation: For a given value 1, no
value is smaller than it and closest
to it in the given array.
Your Task:
You need to complete the given function immediateSmaller() which takes 3 arguments(array arr, N and X) and returns the answer. The printing is done by the driver code.

Expected Time Complexity: O(N).
Expected Auxiliary Space: O(1).

Constraints:
1 <= N <= 103
1 <= arr[i], X <= 104